requirements.txt
aniso8601==7.0.0
Django==2.1.5
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.1.5
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
PyJWT==2.0.0
pytz==2020.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
